requirements.txt
colorama
discord.py
asyncio
re
requests
datetime
